+++ src/import.h
@@ -1,0 +1,12 @@
+#ifndef DIARY_IMPORT_H
+#define DIARY_IMPORT_H
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <regex.h>
+#include <stdbool.h>
+#include "utils.h"
+
+void ics_import(const char* ics_input);
+
+#endif
