+++ src/utils.h
@@ -26,11 +26,10 @@ void update_date(WINDOW* header, struct tm* curs_date);
 char* extract_json_value(const char* json, char* key, bool quoted);
 char* fold(const char* str);
 char* unfold(const char* str);
-char* extract_ical_field(const char* ical, char* key, bool multline);
+char* extract_ical_field(const char* ical, char* key, long* start_pos, bool multline);
 char* expand_path(const char* str);
 char* strrstr(char *haystack, char *needle);
 void fpath(const char* dir, size_t dir_size, const struct tm* date, char** rpath, size_t rpath_size);
-void ics_import(const char* ics_input);
 
 typedef struct
 {
